create authenticate process

+++ routes/tasks.js
@@ -2,9 +2,11 @@ module.exports = app => {
     const Tasks = app.db.models.Tasks;
 
     app.route('/tasks')
+
+        .all(app.auth.authenticate())
         
         .get((req, res) => {
-            Tasks.findAll({})
+            Tasks.findAll({where: {user_id: req.user.id}})
                 .then(result => res.json(result))
                 .catch(error => {
                     res.status(412).json({errorMessage: error.message})
@@ -12,6 +14,7 @@ module.exports = app => {
         })
         
         .post((req, res) => {
+            req.body.user_id  = req.body.id;
             Tasks.create(req.body)
                 .then(result => res.json(result))
                 .catch(error => {
@@ -21,9 +24,11 @@ module.exports = app => {
         });
 
     app.route('/tasks/:id')
+
+        .all(app.auth.authenticate())
         
         .get((req, res) => {
-            Tasks.findOne({where: req.params})
+            Tasks.findOne({where: {id: req.params, user_id: req.user.id} })
                 .then(result => {
                     if (result) {
                         res.json(result);
@@ -37,7 +42,7 @@ module.exports = app => {
         })
         
         .put((req, res) => {
-            Tasks.update(req.body, {where: req.params})
+            Tasks.update(req.body, {where: {id: req.params, user_id: req.user.id} })
                 .then(result => res.sendStatus(204))
                 .catch(error => {
                     res.status(412).json({errorMessage: error.message});
@@ -46,7 +51,7 @@ module.exports = app => {
         })
         
         .delete((req , res) => {
-            Tasks.destroy({where: req.params})
+            Tasks.destroy({where: {id: req.params, user_id: req.user.id} })
             .then(result => res.sendStatus(204))
             .catch(error => {
                 res.status(412).json({errorMessage: error.message});
